Extract migration creation in an util
