+++ packages/twenty-server/src/engine/metadata-modules/object-metadata/object-metadata.service.ts
@@ -24,7 +24,6 @@ import { WorkspaceMigrationRunnerService } from 'src/engine/workspace-manager/wo
 import {
   WorkspaceMigrationColumnActionType,
   WorkspaceMigrationColumnDrop,
-  WorkspaceMigrationCreateComment,
   WorkspaceMigrationTableActionType,
 } from 'src/engine/metadata-modules/workspace-migration/workspace-migration.entity';
 import {
@@ -55,15 +54,15 @@ import {
   createForeignKeyDeterministicUuid,
   createRelationDeterministicUuid,
 } from 'src/engine/workspace-manager/workspace-sync-metadata/utils/create-deterministic-uuid.util';
-import { createWorkspaceMigrationsForCustomObject } from 'src/engine/metadata-modules/object-metadata/utils/create-workspace-migrations-for-custom-object.util';
-import { createWorkspaceMigrationsForRemoteObject } from 'src/engine/metadata-modules/object-metadata/utils/create-workspace-migrations-for-remote-object.util';
+import { createWorkspaceMigrationsForCustomObjectRelations } from 'src/engine/metadata-modules/object-metadata/utils/create-migrations-for-custom-object-relations.util';
+import { createWorkspaceMigrationsForRemoteObjectRelations } from 'src/engine/metadata-modules/object-metadata/utils/create-workspace-migrations-for-remote-object-relations.util';
 import { computeColumnName } from 'src/engine/metadata-modules/field-metadata/utils/compute-column-name.util';
 import { DataSourceEntity } from 'src/engine/metadata-modules/data-source/data-source.entity';
 import { validateObjectMetadataInput } from 'src/engine/metadata-modules/object-metadata/utils/validate-object-metadata-input.util';
 import { mapUdtNameToFieldType } from 'src/engine/metadata-modules/remote-server/remote-table/utils/udt-name-mapper.util';
 import { WorkspaceCacheVersionService } from 'src/engine/metadata-modules/workspace-cache-version/workspace-cache-version.service';
 import { UpdateOneObjectInput } from 'src/engine/metadata-modules/object-metadata/dtos/update-object.input';
-import { buildAlteredCommentOnForeignKeyDeletion } from 'src/engine/metadata-modules/object-metadata/utils/build-altered-comment-on-foreign-key-deletion.util';
+import { createMigrationToAlterCommentOnForeignKeyDeletion } from 'src/engine/metadata-modules/object-metadata/utils/create-migration-for-foreign-key-comment-alteration.util';
 
 import { ObjectMetadataEntity } from './object-metadata.entity';
 
@@ -210,7 +209,10 @@ export class ObjectMetadataService extends TypeOrmQueryService<ObjectMetadataEnt
 
         // for remote objects, we need to update the comment of the foreign key column
         if (objectMetadata.isRemote) {
-          await this.createMigrationToAlterCommentOnForeignKeyDeletion(
+          await createMigrationToAlterCommentOnForeignKeyDeletion(
+            this.dataSourceService,
+            this.typeORMService,
+            this.workspaceMigrationService,
             workspaceId,
             relationToDelete,
           );
@@ -553,7 +555,7 @@ export class ObjectMetadataService extends TypeOrmQueryService<ObjectMetadataEnt
       generateMigrationName(`create-${createdObjectMetadata.nameSingular}`),
       createdObjectMetadata.workspaceId,
       isRemoteObject
-        ? await createWorkspaceMigrationsForRemoteObject(
+        ? await createWorkspaceMigrationsForRemoteObjectRelations(
             createdObjectMetadata,
             activityTargetObjectMetadata,
             attachmentObjectMetadata,
@@ -563,7 +565,7 @@ export class ObjectMetadataService extends TypeOrmQueryService<ObjectMetadataEnt
             objectMetadataInput.primaryKeyColumnType ?? 'uuid',
             workspaceDataSource,
           )
-        : createWorkspaceMigrationsForCustomObject(
+        : createWorkspaceMigrationsForCustomObjectRelations(
             createdObjectMetadata,
             activityTargetObjectMetadata,
             attachmentObjectMetadata,
@@ -968,49 +970,5 @@ export class ObjectMetadataService extends TypeOrmQueryService<ObjectMetadataEnt
     ]);
 
     return { favoriteObjectMetadata };
-  }
-
-  private async createMigrationToAlterCommentOnForeignKeyDeletion(
-    workspaceId: string,
-    relationToDelete: RelationToDelete,
-  ) {
-    const dataSourceMetadata =
-      await this.dataSourceService.getLastDataSourceMetadataFromWorkspaceIdOrFail(
-        workspaceId,
-      );
-
-    const workspaceDataSource =
-      await this.typeORMService.connectToDataSource(dataSourceMetadata);
-
-    const alteredComment = await buildAlteredCommentOnForeignKeyDeletion(
-      relationToDelete.toObjectName,
-      relationToDelete.fromObjectName,
-      dataSourceMetadata.schema,
-      workspaceDataSource,
-    );
-
-    if (alteredComment) {
-      await this.workspaceMigrationService.createCustomMigration(
-        generateMigrationName(
-          `alter-comment-${relationToDelete.fromObjectName}-${relationToDelete.toObjectName}`,
-        ),
-        workspaceId,
-        [
-          {
-            name: computeTableName(
-              relationToDelete.toObjectName,
-              relationToDelete.toObjectMetadataIsCustom,
-            ),
-            action: WorkspaceMigrationTableActionType.ALTER,
-            columns: [
-              {
-                action: WorkspaceMigrationColumnActionType.CREATE_COMMENT,
-                comment: alteredComment,
-              } satisfies WorkspaceMigrationCreateComment,
-            ],
-          },
-        ],
-      );
-    }
   }
 }
